requirements.txt
py-clob-client
python-dotenv
httpx[socks]
feedparser
eth-account
web3
google-api-python-client
google-auth-oauthlib
